requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
channels>=4.0
channels-redis>=4.1
psycopg2-binary>=2.9
django-cors-headers>=4.3
Pillow>=10.0
drf-spectacular>=0.27
redis>=5.0
